Update to use pydantic
